% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{franchise}
\alias{franchise}
\title{franchise data set}
\format{A data.frame with 13 columns
\describe{
 \item{Franchise}{Franchise name}
 \item{Lg}{League}
 \item{From}{Year the team established}
 \item{To}{To year}
 \item{Years}{Years}
 \item{G}{Games}
 \item{W}{Wins}
 \item{L}{Losses}
 \item{Per}{Win Percentage}
 \item{Playoffs}{Playoffs}
 \item{Division}{Division titles}
 \item{Conference}{Conference titles}
 \item{Champions}{Champions}
}}
\source{
Basketball Reference
}
\usage{
franchise
}
\description{
Data set of NBA franchises
}
\keyword{datasets}
