% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFranchise.R
\name{getFranchise}
\alias{getFranchise}
\title{Function for getting Franchises' data: founded year, W/L percentage, number of division/conf/league championship}
\usage{
getFranchise()
}
\description{
Function for getting Franchises' data: founded year, W/L percentage, number of division/conf/league championship
}
