% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHOF.R
\name{getHOF}
\alias{getHOF}
\title{List of Hall of Famers}
\usage{
getHOF()
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item Year
 \item Inductees
 \item Pos.
 \item Achievements
}
}
\description{
Function to get list of Hall of Famers
}
\examples{
\dontrun{
  NBAhof <- getHOF()
  head(NBAhof)
}

}
\seealso{
\url{https://en.wikipedia.org/wiki/List_of_players_in_the_Naismith_Memorial_Basketball_Hall_of_Fame}
}
\author{
Koki Ando
}
