% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaders.R
\name{getLeaders}
\alias{getLeaders}
\title{Stats leaders of a given season}
\usage{
getLeaders(stats_type = c("3PM", "2PM", "3PA", "2PA", "PTS", "AST",
  "BLK", "STL", "FG", "FGA", "FLS", "MP", "TOV", "G", "PF"),
  range = c("Single Season", "Career", "Active"))
}
\arguments{
\item{stats_type}{Stats type you want to fetch information for}

\item{range}{Duration of data. Valid values are Career, Active or Single Season.}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item Rank
 \item Player
 \item X3P
 \item Season
}
}
\description{
Function for getting data of stats leaders of each season
}
\examples{
\dontrun{
 leaders_3pm <- getLeaders(stats_type="3PM", range="Single Season")
 head(leaders_3pm)
}

}
\seealso{
\url{https://www.basketball-reference.com/leaders/}
}
\author{
Koki Ando <koki.25.ando@gmail.com>
}
