% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStandings.R
\name{getStandings}
\alias{getStandings}
\title{Standing Data}
\usage{
getStandings(year, conf = c("East", "West", "All"))
}
\arguments{
\item{conf}{}
}
\value{
This function returns \code{data.fram} including columns:
\itemize{
  \item Team
  \item W
  \item L
  \item W.L.
  \item GB
  \item PW
  \item PL
  \item PS.G
  \item PA.G
}
}
\description{
Function to get historical league standings
}
\seealso{
\url{https://www.basketball-reference.com/leagues/NBA_2019_standings.html}
}
\author{
Koki Ando
}
