% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatsLeader.R
\name{getStatsLeader}
\alias{getStatsLeader}
\title{Stats Leaders}
\usage{
getStatsLeader(stats_type = c("PTS", "AST", "STL", "BLK", "FG2", "FG3"),
  RegularSeason = TRUE)
}
\arguments{
\item{stats_type}{Stats type to fetch information for}

\item{RegularSeason}{Regular season or not. Valid values are TRUE or FALSE}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item Rank
 \item Player
 \item PTS
 \item Season
}
}
\description{
Function for getting stats leader
}
\examples{
\dontrun{
 pts_leader = getStatsLeader(stats_type = "PTS", RegularSeason = TRUE)
 head(pts_leader)
}

}
\seealso{
\url{https://www.basketball-reference.com/leaders/}
}
\author{
Koki Ando
}
