% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatsPerGame.R
\name{getStatsPerGame}
\alias{getStatsPerGame}
\title{Stats Per Game for a given player}
\usage{
getStatsPerGame(Player, season, span = 1)
}
\arguments{
\item{Player}{The name of the player to fetch information for}

\item{season}{NBA season number}

\item{span}{Duration of data}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item G
 \item Date
 \item Age
 \item Tm
 \item Var.6
 \item Opp
 \item Var.8
 \item GS
 \item MP
 \item FG
 \item FGA
 \item FG.
 \item X3P
 \item X3PA
 \item X3P.
 \item FT
 \item FTA
 \item FT.
 \item ORB
 \item DRB
 \item TRB
 \item AST
 \item STL
 \item BLK
 \item TOV
 \item PF
 \item PTS
 \item GmSc
 \item X...
}
}
\description{
Function for getting given player's stas per game
}
\examples{
\dontrun{
getStatsPerGame(Player="Kobe Bryant", season="2008")
}

}
\author{
Koki Ando
}
