% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatsSummary.R
\name{getStatsSummary}
\alias{getStatsSummary}
\title{Function for getting given player's career summary stats}
\usage{
getStatsSummary(Name)
}
\arguments{
\item{Name}{palyers name}
}
\description{
Function for getting given player's career summary stats
}
\examples{

}
