% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HallOfFame}
\alias{HallOfFame}
\title{HallOfFame data set}
\format{A data.frame with 4 columns
\describe{
 \item{Year}{Year}
 \item{Inductee}{Name of inductee}
 \item{Pos}{Position}
 \item{Achievements}{Achievements}
}}
\source{
wikipedia
}
\usage{
HallOfFame
}
\description{
Data set of players in the Naismith Memorial Basketball Hall of Fame
}
\keyword{datasets}
