% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBAloveR.R
\docType{package}
\name{NBAloveR}
\alias{NBAloveR}
\alias{NBAloveR-package}
\title{NBaloveR: Provides data access to Basketball data}
\description{
R package that provides an interdace to online basketball data. This package allows R users to download
publicly available data on basketball website such as Basketball Reference <>, NBA Draft<>
}
\author{
Koki Ando \email{koki.25.ando@gmail.com}
}
