% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{player_data}
\alias{player_data}
\title{player_data data set}
\format{
A data.frame with 12 columns
\describe{
 \item{Player}{PLayer name}
 \item{Pos}{Position}
 \item{HT}{Height}
 \item{WT}{Weight}
 \item{Age}{Age}
 \item{Teams}{Teams}
 \item{GP}{Games Played}
 \item{YOS}{Year of season}
 \item{Pre-Draft Team}{Pre Draft Team}
 \item{Draft Status}{Draft Status}
 \item{Nationality}{Nationality}
 \item{url}{reference url}
}
}
\source{
RealGM
}
\usage{
player_data
}
\description{
Data set of players information. from https://basketball.realgm.com
}
\keyword{datasets}
