% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{players}
\alias{players}
\title{players data set}
\format{
A data.frame with 13 columns
\describe{
 \item{Player}{Name of Player}
 \item{POS}{Position of given player}
 \item{HT}{Height}
 \item{WT}{Weight}
 \item{Age}{Age}
 \item{Teams}{Team(s) the given player belongs to}
 \item{GP}{Games played}
 \item{YOS}{Year of seasons}
 \item{PreDraftTeam}{Pre-draft team}
 \item{Draft.Status}{Draft status}
 \item{Nationality}{Nationality}
 \item{Year}{Year}
 \item{RookieYear}{Rookie year}
}
}
\source{
Basketball Real GM
}
\usage{
players
}
\description{
Data set of all NBA players
}
\keyword{datasets}
