% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonSchedule.R
\name{seasonSchedule}
\alias{seasonSchedule}
\title{Results of each game for a given team in give season}
\usage{
seasonSchedule(Team, year)
}
\arguments{
\item{year}{}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item G
 \item Date
 \item Start..ET.
 \item Var.6
 \item Opponent
 \item Var.8
 \item Var.9
 \item Tm
 \item Opp
 \item W
 \item L
 \item Streak
}
}
\description{
Function to get result of games
}
\examples{
\dontrun{
 bos08 <- seasonSchedule(Team = "bos", year = 2008)
 head(bos08)
}

}
\seealso{
\url{https://www.basketball-reference.com/teams/}
}
\author{
Koki Ando
}
