% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDraftResult.R
\name{getDraftResult}
\alias{getDraftResult}
\title{Draft Result of a given year}
\usage{
getDraftResult(year)
}
\arguments{
\item{year}{}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item #
 \item Team
 \item Player
 \item School
}
}
\description{
Function to get draft result
}
\seealso{
\url{http://www.nbadraft.net}
}
\author{
Koki Ando
}
