% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDraftResult.R
\name{getDraftResult}
\alias{getDraftResult}
\title{Draft Result of a given year}
\usage{
getDraftResult(year)
}
\arguments{
\item{year}{Year number}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item #
 \item Team
 \item Player
 \item School
}
}
\description{
Function to get draft result
}
\examples{
\dontrun{
  draft09 <- getDraftResult(year = 1999)
  draft09 <- getDraftResult(year = 2006)
  head(draft09)
}

}
\seealso{
\url{http://www.nbadraft.net}
}
\author{
Koki Ando <koki.25.ando@gmail.com>
}
