% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{getMatchups}
\alias{getMatchups}
\title{Season Matchups}
\usage{
getMatchups(team_code, season)
}
\arguments{
\item{team_code}{Team code consisting 3 characters}

\item{season}{Season number}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item {"G"}{Games}
 \item {"Date"}{Date}
 \item {"StartTime"}{Time-zone ET}
 \item {"Opponent"}{Opponent Team}
 \item {"Result"}{Game Result}
 \item {"OT"}{Over Time}
 \item {"Tm"}{Points}
 \item {"Opp"}{Opponent Points}
 \item {"W"}{Wins}
 \item {"L"}{Losses}
 \item {"Streak"}{Streaks}
}
}
\description{
Season Matchups and results of a given team
}
\examples{
\dontrun{
 bos19 <- getMatchups(team_code = "bos", season = 2019)
 head(bos19)
}

}
\seealso{
\url{https://www.basketball-reference.com/teams/}
}
\author{
Koki Ando
}
