% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatsPerGame.R
\name{getStatsPerGame}
\alias{getStatsPerGame}
\title{Stats Per Game for a given player}
\usage{
getStatsPerGame(Player, season, span = 1)
}
\arguments{
\item{Player}{The name of the player to fetch information for}

\item{season}{NBA season number}

\item{span}{Duration of data}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item G
 \item Date
 \item Age
 \item Tm
 \item Home
 \item Opp
 \item GS
 \item MP
 \item FG
 \item FGA
 \item FGP
 \item 3PM
 \item 3PA
 \item 3PP
 \item FT
 \item FTA
 \item FTP
 \item ORB
 \item DRB
 \item TRB
 \item AST
 \item STL
 \item BLK
 \item TOV
 \item PF
 \item PTS
 \item GameScore
 \item PlusMinus
}
}
\description{
Function for getting given player's stats per game
}
\examples{
\dontrun{
kobe08 <- getStatsPerGame(Player="Kobe Bryant", season="2008")
head(kobe08)
}

}
\author{
Koki Ando <koki.25.ando@gmail.com>
}
