% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{getStatsSummary}
\alias{getStatsSummary}
\title{Career Stats Summary}
\usage{
getStatsSummary(player)
}
\arguments{
\item{player}{Player name}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item Season
 \item Age
 \item Tm
 \item Lg
 \item Pos
 \item G
 \item GS
 \item MP
 \item FG
 \item FG%
 \item 3P
 \item 3PA
 \item 3P%
 \item 2P
 \item 2PA
 \item 2P%
 \item eFG%
 \item FT
 \item FTA
 \item FT%
 \item ORB
 \item DRB
 \item TRB
 \item AST
 \item STL
 \item BLK
 \item TOV
 \item PF
 \item PTS
}
}
\description{
Career stats summary of a given player
}
\examples{
\dontrun{
 kobe_summary  = getStatsSummary(player = "Kobe Bryant")
 head(kobe_summary)
}

}
\seealso{
\url{https://www.basketball-reference.com}
}
\author{
Koki Ando <koki.25.ando@gmail.com>

Koki Ando <koki.25.ando@gmail.com>
}
