% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatsSummary.R
\name{getStatsSummary}
\alias{getStatsSummary}
\title{Career summary stats for a given player}
\usage{
getStatsSummary(Name)
}
\arguments{
\item{Name}{The name of the player to fetch information for}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item Season
 \item Age
 \item Tm
 \item Lg
 \item Pos
 \item G
 \item GS
 \item MP
 \item FG
 \item FG%
 \item 3P
 \item 3PA
 \item 3P%
 \item 2P
 \item 2PA
 \item 2P%
 \item eFG%
 \item FT
 \item FTA
 \item FT%
 \item ORB
 \item DRB
 \item TRB
 \item AST
 \item STL
 \item BLK
 \item TOV
 \item PF
 \item PTS
}
}
\description{
Function for getting given player's career summary stats
}
\examples{
\dontrun{
<<<<<<< HEAD
  getStatsSummary(Name = "Kobe Bryant")
=======
  getStatsSummary("Kobe Bryant")
>>>>>>> a326d2ccc47dec97d5501539f2a2846eed11a994
}

}
\seealso{
\url{https://www.basketball-reference.com}
}
\author{
Koki Ando <koki.25.ando@gmail.com>
}
