% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTeamSalary.R
\name{getTeamSalary}
\alias{getTeamSalary}
\title{Salary data for a given team}
\usage{
getTeamSalary(team_code)
}
\arguments{
\item{team_code}{Team code consisting of 3 characters to fetch information for}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item Player
 \item Season
 \item Signed Using
 \item Guaranteed
}
}
\description{
Function that allows you to get salary data of each team
}
\examples{
\dontrun{
 salary <- getTeamSalary(team_code = "gsw") # Extract salary dataset of Golden State Warriors' players
 head(salary)
}

}
\seealso{
\url{https://www.basketball-reference.com/contracts/GSW.html}
}
\author{
Koki Ando
}
