% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBAloveR.R
\docType{package}
\name{NBAloveR}
\alias{NBAloveR}
\alias{NBAloveR-package}
\title{NBaloveR: Provides data access to basketball data}
\description{
R package that provides an interface to online basketball data. This package allows R users to download
publicly available data on basketball.
}
\author{
Koki Ando \email{koki.25.ando@gmail.com}
}
