% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsCompare.R
\name{statsCompare}
\alias{statsCompare}
\title{Visualization of Stats comparison}
\usage{
statsCompare(player_list = c(), Age = FALSE)
}
\arguments{
\item{player_list}{List of names of players.}

\item{Age}{Age valid values are TRUE or FALSE}
}
\value{
This function returns a point and line plot showing transitions of PPG stats of given players.
}
\description{
Easy stats comparison function, which also includes simple line and point plots
}
\examples{
\dontrun{
  statsCompare(c("Kevin Durant", "Russel Westbrook", "James Harden"), Age=FALSE)
}

}
\author{
Koki Ando <koki.25.ando@gmail.com>
}
