% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsCompare.R
\name{statsCompare}
\alias{statsCompare}
\title{Line plot for Stats comparison}
\usage{
statsCompare(player_list = c(), Age = FALSE)
}
\arguments{
\item{Age}{}
}
\description{
Easy stats comparison function, which also includes simple line plots
}
\examples{
\dontrun{
  statsCompare(c("Kobe Bryant", "Allen Iverson", "Paul Pierce"), Age=TRUE)
}

}
\author{
Koki Ando
}
